Fix long-to-int truncation in FlowTimeDemo delay calculation (#1271)
Use long instead of int for demandDelay and stepToGet to match
getCurrentStep() return type and avoid silent truncation under
extreme parameterization.

+++ courant-demos/src/main/java/systems/courant/sd/demo/FlowTimeDemo.java
@@ -66,8 +66,8 @@ public class FlowTimeDemo {
         Flow demand = Flows.linearGrowth("New Orders", DAY, wip, newOrdersPerDay);
 
         Flow throughput = Flow.create("Delivered Reports", DAY, () -> {
-            int demandDelay = Math.max(0, (int) Math.round(tat.getValue()));
-            int stepToGet = Math.max(0, (int) (sim.getCurrentStep() - demandDelay));
+            long demandDelay = Math.max(0, Math.round(tat.getValue()));
+            long stepToGet = Math.max(0, sim.getCurrentStep() - demandDelay);
             double demandPlusDelay = demand.getHistoryAtTimeStep(stepToGet);
             return new Quantity(Math.min(capacity, demandPlusDelay), ItemUnits.THING);
         });
